chore: Consumed native types in tsconfig.json
