+++ docs/ember-intl/tsconfig.json
@@ -11,7 +11,10 @@
       "docs-app-for-ember-intl/*": ["app/*"],
       "*": ["types/*"]
     },
-    "skipLibCheck": true
+    "skipLibCheck": true,
+    "types": [
+      "ember-source/types"
+    ]
   },
   "include": [
     "app/**/*",
