+++ docs/ember-intl/types/ember-cli-addon-docs.ts
