+++ docs/my-app-with-fallbacks/types/my-v1-addon.ts
