+++ docs/my-app-with-namespace-from-folders/types/my-v1-addon.ts
