+++ docs/my-app/types/my-v1-addon.ts
