+++ docs/my-classic-app-with-lazy-loaded-translations/types/my-v1-addon.ts
