+++ docs/my-classic-app/types/my-v1-addon.ts
