+++ docs/my-v1-engine/types/global.d.ts
@@ -1,6 +1,5 @@
 import '@glint/environment-ember-loose';
 import '@glint/environment-ember-template-imports';
-import 'ember-source/types';
 
 import type EmberIntlRegistry from 'ember-intl/template-registry';
 import type EmberPageTitleRegistry from 'ember-page-title/template-registry';
