+++ packages/ember-intl/tsconfig.json
@@ -10,7 +10,10 @@
       "ember-intl/test-support": ["addon-test-support"],
       "ember-intl/test-support/*": ["addon-test-support/*"],
       "*": ["types/*"]
-    }
+    },
+    "types": [
+      "ember-source/types"
+    ]
   },
   "include": [
     "addon/**/*",
