--- packages/ember-intl/types/global.d.ts
@@ -1,6 +1,0 @@
-// Types for compiled templates
-declare module 'ember-intl/templates/*' {
-  import type { TemplateFactory } from 'ember-cli-htmlbars';
-  const tmpl: TemplateFactory;
-  export default tmpl;
-}
